add new tool - text editor
