+++ pyproject.toml
@@ -26,6 +26,7 @@ dependencies = [
     "pytest-asyncio>=1.0.0",
     "python-dotenv>=1.1.0",
     "sqlalchemy~=2.0",
+    "strands-agents-tools>=0.2.0",
     "uvicorn>=0.34.2",
     "wikipedia>=1.4.0",
 ]
